machine_learning/embeddings/find_similar_images/requirements.txt
torch
torchvision
transformers
faiss-cpu
pillow